Fix a license copy-and-pasto.
The Twitter-API-PHP package is actually under the MIT License (see
https://github.com/J7mbo/twitter-api-php).  That's perfectly
compatible with GNU AGPL, so it's no problem for Checkbook; we just
have to keep the original license.  The license here probably just got
accidentally caught in a license header sweep done across the whole
checkbook_alerts module before it was imported.  This commit restores
the MIT License metadata as per the upstream package.

Also, I'll note in this commit message because there isn't anywhere
else to note it: in commit da1f87b0e1, which imported this package,
the "CURLOPT_TIMEOUT => 10" was already removed from the $options
array in performRequest() in TwitterAPIExchange.php -- thus what was
imported here differs slightly from what's upstream.  This is not a
problem, license-wise or code-wise, but being a fanatic about keeping
records, I'm mentioning it here, since no 'git diff' in our tree's
history will ever show it.