Remove obsolete 2013 and 2014 demo data files
The 2013 and 2014 data files are obsolete now, and rather large.
Remove them, leaving just the 2015 demo data, which is the latest.

Also, make both 2015 data files use the .7z format (with the p7zip
package).  It looks like in commit e5a1f2af3 we started using the .7z
format, presumably because it has a better compression algorithm, and
without .7z the main demo DB data file is actually too large for
Github (> 100 MB).  For the sake of consistency, I put the ogent demo
DB in .7z format as well.


Former-commit-id: c1a077213da4fe90b15b8fd0c8870277076022f0 [formerly c1a077213da4fe90b15b8fd0c8870277076022f0 [formerly c1a077213da4fe90b15b8fd0c8870277076022f0 [formerly 0cc09625600d4eda908e085a9906bb03543995a2]]]
Former-commit-id: fa9f1c000b7a9d0d6ea43436a6ebd4fd47392a7e
Former-commit-id: 5af0ce10755df9b7b819d37f5f277a6812a46eea
Former-commit-id: 7ceb8a0dd565b6675b6cbde0f59c6d42e3423bd4