Remove obsolete 2013 and 2014 demo data files
The 2013 and 2014 data files are obsolete now, and rather large.
Remove them, leaving just the 2015 demo data, which is the latest.

Also, make both 2015 data files use the .7z format (with the p7zip
package).  It looks like in commit e5a1f2af3 we started using the .7z
format, presumably because it has a better compression algorithm, and
without .7z the main demo DB data file is actually too large for
Github (> 100 MB).  For the sake of consistency, I put the ogent demo
DB in .7z format as well.