Removed 2013 and 2014 data files, as they're unnecessary (and rather large). The 2015 demo data makes the previous versions obsolete.
Also, made both 2015 data files use the .7z format (with the p7zip
package). It looks like someone switched to the .7z format because
it has a better compression algorithm. Without .7z, the main demo
DB is actually too large for Github (> 100 MB). For the sake of
consistency, I put the ogent demo DB in .7z format as well.