Hack to include javascript and show charts
Manually include javascript libraries so that charts show up on pages
like localhost/spending_landing/yeartype/B/year/113#.  This is not how
this should be done, and continues to have problems where certain
javascript is not loaded.  The next step is to figure out why Drupal
isn't loading the js libraries as it should be, and to fix that -- and
then probably to revert this commit.

Note that some of the elements of $node that were being checked don't
seem to exist (e.g. $node->widgetConfig->chartConfig->title->text was
always empty, even though $node->widgetConfig->chartTitle had a value).
I'm not sure if that's a problem with my local copy or if they've been
deprecated.

I set `var chart` equal to null before setting it to a new Highcharts
Chart object just to avoid having any problem with initializing
Highcharts twice (if it somehow was initialized before being loaded in
this file).

Plenty of things are still broken in this state, though my local
deployment *mostly* works now.  The Filter dropdown doesn't work, the
citywide agencies and other government entities lists are empty, and the
search doesn't run.