Fix agency menu display: #56
Show a "missing agencies" message when no agencies are found in the
data, instead of a broken menu.  Add a conditional to check whether
there are agencies.  If none are found, construct an element to make the
menu render properly.  Fill it with a message stating that no agencies
were found.

With the current sample data, the landing page and related pages did not
display the Citywide Agencies and Other Government Entities menu items
correctly.  This was because there were no city agencies to display, so
the loops in agencies_list.tpl.php were never entered.

A more extensive fix for this issue would be to update the sample data
to include citywide agencies everywhere they are needed.

You may want to view this commit with the -w flag, since adding the
conditional changed the indentation of several lines.