Add Travis CI to this repository
Create a .travis.yml file as per the basic Travis instructions.  The
pom.xml file is not in the root of the repo, so follow the [guidance
here](http://blog.javafortesters.com/2015/07/using-travis-ciorg-for-checking-code-on.html)
about pointing Travis to a different location for the test base.

Add a link to the INSTALL file about setting up Travis for different
forks.

Once this is set up correctly, Travis will run `mvn test` on every
commit coming into the repository.